Update type definitions in header files for consistency and accuracy

+++ include/render.h
@@ -6,7 +6,7 @@
 /*   By: descamil <descamil@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2025/04/04 12:33:49 by descamil          #+#    #+#             */
-/*   Updated: 2025/04/08 16:11:11 by descamil         ###   ########.fr       */
+/*   Updated: 2025/04/09 19:31:14 by descamil         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -17,9 +17,9 @@
 
 // ft_maths.c
 t_vec3	ft_normalice(t_vec3 vec);
-float	ft_dot(t_vec3 a, t_vec3 b);
+double	ft_dot(t_vec3 a, t_vec3 b);
 t_vec3	ft_cross(t_vec3 a, t_vec3 b);
-t_vec3	ft_scale(t_vec3 value, float scale);
+t_vec3	ft_scale(t_vec3 value, double scale);
 
 // ft_render.c
 void	ft_create_render(t_mlx *data, t_image *image);
